pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# All the following settings are optional:
where = ["."]  # ["."] by default
include = ["*"]  # ["*"] by default
exclude = ["inputs*", "outputs*", "docs*"]
namespaces = true  # true by default

[tool.setuptools.package-data]
"zigzag" = ["py.typed"]

[project]
name = "zigzag-dse"
version = "3.3.0"
description = "ZigZag - Deep Learning Hardware Design Space Exploration"
readme = "README.md"
authors = [{ name = "Arne Symons", email = "arne.symons@kuleuven.be" }, { name = "Linyan Mei", email = "linyan.mei@kuleuven.be" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
]
keywords = ["zigzag", "dse", "design-space-exploration", "machine-learning", "deep-learning", "mapping"]
dependencies = [
    'numpy',
    'networkx',
    'sympy',
    'matplotlib',
    'onnx',
    'tqdm',
    'multiprocessing_on_dill',
    'pyyaml',
    'cerberus',
    'seaborn'
]
requires-python = ">=3.11"


[project.optional-dependencies]
dev = ["bumpver", "pip-tools", "build", "twine", 'pytest', 'typeguard', 'ruff']

[project.urls]
Homepage = "https://github.com/KULeuven-MICAS/zigzag"

[project.scripts]
realpython = "zigzag.__main__:main"

[tool.bumpver]
current_version = "3.3.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"', 
    'version = "{version}"'
]
"zigzag/__init__.py" = ["{version}"]


[tool.ruff]
line-length = 120

[tool.black]
line-length = 120

[tool.pyright]
include = ["zigzag"]
typeCheckingMode="strict"
pythonVersion = "3.11"
stubPath = "./stubs/stubs"
useLibraryCodeForTypes = true
reportInvalidTypeForm = "none"
reportMissingModuleSource = "none"
reportMissingImports = "none"
reportAttributeAccessIssue ="none" # False positives for some packages
